Code Refactoring
- move 'backend/configs/' to 'backend/config/'
- move `verifySession` to `middleware/`
- move blockchain routes to 'backend/routes/blockchain.js'
- move users routes to 'backend/routes/users.js'
-move 'backend/fabric' to 'backend/services'
- rename all userID variables to "userID"
- update .gitignore
