+++ .gitignore
@@ -102,4 +102,3 @@ dist
 
 # TernJS port file
 .tern-port
-backend/configs/config.json
