+++ backend/index.js
@@ -1,14 +1,8 @@
 'use strict'
 
-const path = require('path');
-const fs = require('fs');
-const util = require('util');
-const jwt = require('jsonwebtoken');
 const bodyParser = require('body-parser');
 
-const { User } = require('./models');
 const { auth } = require('./middleware/auth');
-const pollRouter = require('./routes/poll');
 
 const express = require('express');
 const app = express();
@@ -16,12 +10,10 @@ const app = express();
 // Setup mongoose
 require("./startup/mongoose")();
 
-// Blockchain Network
-let network = require('./fabric/network.js');
-const configPath = path.join(process.cwd(), './configs/config.json');
-const configJSON = fs.readFileSync(configPath, 'utf8');
-const config = JSON.parse(configJSON);
-const appAdmin = config.appAdmin;
+// Router
+const pollRouter = require('./routes/poll');
+const usersRouter = require('./routes/users');
+const blockchainRouter = require('./routes/blockchain');
 
 /* MIDDLEWARE  */
 app.use(bodyParser.json());
@@ -38,144 +30,12 @@ app.use(function (req, res, next) {
     next();
 });
 
-// Verify Refresh Token Middleware (which will be verifying the session)
-let verifySession = (req, res, next) => {
-    // grab the refresh token from the request header
-    let refreshToken = req.header('x-refresh-token');
-
-    // grab the _id from the request header
-    let _id = req.header('_id');
-
-    User.findByIdAndToken(_id, refreshToken).then((user) => {
-        if (!user) {
-            // user couldn't be found
-            return Promise.reject({
-                'error': 'User not found. Make sure that the refresh token and user id are correct'
-            });
-        }
-
-
-        // if the code reaches here - the user was found
-        // therefore the refresh token exists in the database - but we still have to check if it has expired or not
-
-        req.userID = user._id;
-        req.userObject = user;
-        req.refreshToken = refreshToken;
-
-        let isSessionValid = false;
-
-        user.sessions.forEach((session) => {
-            if (session.token === refreshToken) {
-                // check if the session has expired
-                if (User.hasRefreshTokenExpired(session.expiresAt) === false) {
-                    // refresh token has not expired
-                    isSessionValid = true;
-                }
-            }
-        });
-
-        if (isSessionValid) {
-            // the session is VALID - call next() to continue with processing this web request
-            next();
-        } else {
-            // the session is not valid
-            return Promise.reject({
-                'error': 'Refresh token has expired or the session is invalid'
-            })
-        }
-
-    }).catch((e) => {
-        res.status(401).send(e);
-    })
-}
 /* END MIDDLEWARE  */
 
 /* ROUTE HANDLERS */
 app.use('/poll',auth, pollRouter);
-
-/* USER ROUTES */
-
-/**
- * POST /users
- * Purpose: Sign up
- */
-app.post('/users', (req, res) => {
-    // User sign up
-
-    let body = req.body;
-    let newUser = new User(body);
-
-    newUser.save().then(() => {
-        return newUser.createSession();
-    }).then((refreshToken) => {
-        // Session created successfully - refreshToken returned.
-        // now we geneate an access auth token for the user
-
-        return newUser.generateAccessAuthToken().then((accessToken) => {
-            // access auth token generated successfully, now we return an object containing the auth tokens
-            return { accessToken, refreshToken }
-        });
-    }).then(async (authTokens) => {
-        // Sign up user in Blockchain and add to wallet
-        await network.registerUser(newUser._id.toString());
-
-        // Now we construct and send the response to the user with their auth tokens in the header and the user object in the body
-        res
-            .header('x-refresh-token', authTokens.refreshToken)
-            .header('x-access-token', authTokens.accessToken)
-            .send(newUser);
-    }).catch((e) => {
-        res.status(400).send(e);
-    })
-})
-
-
-/**
- * POST /users/login
- * Purpose: Login
- */
-app.post('/users/login', (req, res) => {
-    let email = req.body.email;
-    let password = req.body.password;
-
-    User.findByCredentials(email, password).then((user) => {
-        return user.createSession().then((refreshToken) => {
-            // Session created successfully - refreshToken returned.
-            // now we geneate an access auth token for the user
-
-            return user.generateAccessAuthToken().then((accessToken) => {
-                // access auth token generated successfully, now we return an object containing the auth tokens
-                return { accessToken, refreshToken }
-            });
-        }).then(async (authTokens) => {
-            // Now we construct and send the response to the user with their auth tokens in the header and the user object in the body
-            // This is how to connect to Blockchain network with userId
-            let connection = await network.connectToNetwork(user._id.toString());
-            
-            res
-                .header('x-refresh-token', authTokens.refreshToken)
-                .header('x-access-token', authTokens.accessToken)
-                .send(user);
-        })
-    }).catch((e) => {
-        res.status(400).send(e);
-    });
-})
-
-
-/**
- * GET /users/me/access-token
- * Purpose: generates and returns an access token
- */
-app.get('/users/me/access-token', verifySession, (req, res) => {
-    // we know that the user/caller is authenticated and we have the userID and user object available to us
-    req.userObject.generateAccessAuthToken().then((accessToken) => {
-        res.header('x-access-token', accessToken).send({ accessToken });
-    }).catch((e) => {
-        res.status(400).send(e);
-    });
-})
-
+app.use('/users', usersRouter);
+app.use('/blockchain', blockchainRouter);
 
 /**
  * POST /poll
@@ -185,32 +45,6 @@ app.post('/poll', (req, res) => {
     console.log(req.body);
     res.send({ 'message': 'updated successfully' });
 })
-
-
-/* HELPER METHODS */
-let deleteTasksFromList = (_listId) => {
-    Task.deleteMany({
-        _listId
-    }).then(() => {
-        console.log("Tasks from " + _listId + " were deleted!");
-    })
-}
-
-
-/* Blockchain ROUTES */
-
-/**
- * GET /blockchain/queryAll
- * Purpose: query all assets in world state
- */
-app.get('/blockchain/queryAll', async (req, res) => {
-    let connection = await network.connectToNetwork(appAdmin);
-    let response = await network.invoke(connection, true, 'queryAll', '');
-    let parsedResponse = await JSON.parse(response);
-    res.send(parsedResponse);
-});
-
-/* End of Blockchain ROUTES */
 
 app.listen(3000, () => {
     console.log("Server is listening on port 3000");
